mint contract - support short position
