+++ artifacts/checksums.txt
@@ -1,8 +1,8 @@
 7c274501ed67880fb3dbe12e943783b83e384677f88c71259f4eba65bce833aa  mirror_collector.wasm
 a424e2479c9999a57e01a097ea9dfe5606cac9dfc876d894ab2e3baa882a7c37  mirror_community.wasm
-519153926d5275a543931f411205a71430f2744ac3bea223dc8a24247d563ee2  mirror_factory.wasm
+157b1d3313f7e3e36d434613c078f57f26c39ad5a3108bc09290c11063da3e2b  mirror_factory.wasm
 97af47b31d996d648ef924c2c3028cd04b07132a59a8d9571e886e9842aec7b7  mirror_gov.wasm
 f41f717d3885e96d180a29f2b3d4f83080f973cd1e5bda1f59ff5420789de5f1  mirror_limit_order.wasm
-4f59f4db25feba6338ef524086c240fb92e9ccbc699ca4205ad7c19114c4907f  mirror_mint.wasm
-e8803e73a12cbc10da0bec94d118db57f3b67fb903825306ec130452f0412e95  mirror_oracle.wasm
-6b232383a3cd398cbc1aba4c65cacad93aa308fe0a2eb07f1e92a3dee1566181  mirror_staking.wasm
+b5fdb00f776c3f517f4f377c5db377d60725b775bc8792c6a361d8e632b3a5db  mirror_mint.wasm
+618464352735de7c0d44e6143d44d1427a775917e056b0a1c664768f0c4e2f88  mirror_oracle.wasm
+8a9fc6b51729f33e51342373655b9dde1252bcbd60b51eb3724dae8fe137a5fa  mirror_staking.wasm
