+++ contracts/mirror_mint/Cargo.toml
@@ -29,12 +29,9 @@ incremental = false
 overflow-checks = true
 
 [features]
-default = ["cranelift"]
 # for quicker tests, cargo test --lib
 # for more explicit tests, cargo test --features=backtraces
-backtraces = ["cosmwasm-std/backtraces", "cosmwasm-vm/backtraces"]
-cranelift = ["cosmwasm-vm/default-cranelift"]
-singlepass = ["cosmwasm-vm/default-singlepass"]
+backtraces = ["cosmwasm-std/backtraces"]
 
 [dependencies]
 cw20 = { version = "0.2" }
@@ -47,5 +44,4 @@ serde = { version = "1.0.103", default-features = false, features = ["derive"] }
 
 [dev-dependencies]
 terra-cosmwasm = { version = "1.2.2" }
-cosmwasm-vm = { version = "0.10.1", default-features = false, features = ["iterator"] }
 cosmwasm-schema = "0.10.1"
