+++ contracts/mirror_mint/src/lib.rs
@@ -1,11 +1,18 @@
 pub mod contract;
-pub mod math;
-pub mod migration;
-pub mod querier;
-pub mod state;
 
-#[cfg(test)]
-mod testing;
+mod asserts;
+mod math;
+mod migration;
+mod positions;
+mod querier;
+mod state;
+
+// Testing files
+mod contract_test;
+mod migrated_asset_test;
+mod positions_test;
+mod pre_ipo_test;
+mod short_test;
 
 #[cfg(test)]
 mod mock_querier;
