+++ contracts/mirror_mint/src/migration.rs
@@ -2,22 +2,47 @@ use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 use cosmwasm_std::{CanonicalAddr, Decimal, Order, StdResult, Storage};
-use cosmwasm_storage::{ReadonlyBucket};
+use cosmwasm_storage::{singleton_read, ReadonlyBucket};
 
-use crate::state::{store_asset_config, AssetConfig};
+use crate::state::{
+    store_asset_config, store_config, AssetConfig, Config, KEY_CONFIG, PREFIX_ASSET_CONFIG,
+};
 
-#[cfg(test)]
-use cosmwasm_std::{Api, HumanAddr};
-#[cfg(test)]
-use cosmwasm_storage::Bucket;
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+pub struct LegacyConfig {
+    pub owner: CanonicalAddr,
+    pub oracle: CanonicalAddr,
+    pub collector: CanonicalAddr,
+    pub base_denom: String,
+    pub token_code_id: u64,
+    pub protocol_fee_rate: Decimal,
+}
 
-static PREFIX_ASSET_CONFIG: &[u8] = b"asset_config";
+fn read_legacy_config<S: Storage>(storage: &S) -> StdResult<LegacyConfig> {
+    singleton_read(storage, KEY_CONFIG).load()
+}
 
-#[cfg(test)]
-pub fn asset_config_old_store<'a, S: Storage>(
-    storage: &'a mut S,
-) -> Bucket<'a, S, LegacyAssetConfig> {
-    Bucket::new(PREFIX_ASSET_CONFIG, storage)
+pub fn migrate_config<S: Storage>(
+    storage: &mut S,
+    staking: CanonicalAddr,
+    terraswap_factory: CanonicalAddr,
+) -> StdResult<()> {
+    let legacy_config: LegacyConfig = read_legacy_config(storage)?;
+    store_config(
+        storage,
+        &Config {
+            staking: staking,
+            terraswap_factory,
+            owner: legacy_config.owner,
+            oracle: legacy_config.oracle,
+            collector: legacy_config.collector,
+            base_denom: legacy_config.base_denom,
+            token_code_id: legacy_config.token_code_id,
+            protocol_fee_rate: legacy_config.protocol_fee_rate,
+        },
+    )?;
+
+    Ok(())
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
@@ -63,8 +88,81 @@ pub fn migrate_asset_configs<S: Storage>(storage: &mut S) -> StdResult<()> {
 #[cfg(test)]
 mod migrate_tests {
     use super::*;
-    use crate::state::read_asset_config;
+    use crate::state::{read_asset_config, read_config};
     use cosmwasm_std::testing::mock_dependencies;
+    use cosmwasm_std::{Api, HumanAddr};
+    use cosmwasm_storage::{singleton, Bucket};
+
+    pub fn asset_config_old_store<'a, S: Storage>(
+        storage: &'a mut S,
+    ) -> Bucket<'a, S, LegacyAssetConfig> {
+        Bucket::new(PREFIX_ASSET_CONFIG, storage)
+    }
+
+    pub fn store_legacy_config<S: Storage>(
+        storage: &mut S,
+        config: &LegacyConfig,
+    ) -> StdResult<()> {
+        singleton(storage, KEY_CONFIG).save(config)
+    }
+
+    #[test]
+    fn test_config_migration() {
+        let mut deps = mock_dependencies(20, &[]);
+
+        let owner = deps
+            .api
+            .canonical_address(&HumanAddr::from("owner0000"))
+            .unwrap();
+        let oracle = deps
+            .api
+            .canonical_address(&HumanAddr::from("oracle0000"))
+            .unwrap();
+        let collector = deps
+            .api
+            .canonical_address(&HumanAddr::from("collector0000"))
+            .unwrap();
+        let staking = deps
+            .api
+            .canonical_address(&HumanAddr::from("staking0000"))
+            .unwrap();
+        let terraswap_factory = deps
+            .api
+            .canonical_address(&HumanAddr::from("terraswap_factory"))
+            .unwrap();
+        store_legacy_config(
+            &mut deps.storage,
+            &LegacyConfig {
+                owner: owner.clone(),
+                oracle: oracle.clone(),
+                collector: collector.clone(),
+                base_denom: "uusd".to_string(),
+                token_code_id: 10,
+                protocol_fee_rate: Decimal::percent(1),
+            },
+        )
+        .unwrap();
+
+        migrate_config(
+            &mut deps.storage,
+            staking.clone(),
+            terraswap_factory.clone(),
+        )
+        .unwrap();
+        assert_eq!(
+            read_config(&deps.storage).unwrap(),
+            Config {
+                owner,
+                oracle,
+                staking,
+                collector,
+                terraswap_factory,
+                base_denom: "uusd".to_string(),
+                token_code_id: 10,
+                protocol_fee_rate: Decimal::percent(1),
+            }
+        );
+    }
 
     #[test]
     fn test_asset_config_migration() {
