--- contracts/mirror_mint/tests/integration.rs
@@ -1,128 +1,0 @@
-//! This integration test tries to run and call the generated wasm.
-//! It depends on a Wasm build being available, which you can create with `cargo wasm`.
-//! Then running `cargo integration-test` will validate we can properly call into that generated Wasm.
-//!
-//! You can easily convert unit tests to integration tests as follows:
-//! 1. Copy them over verbatim
-//! 2. Then change
-//!      let mut deps = mock_dependencies(20, &[]);
-//!    to
-//!      let mut deps = mock_instance(WASM, &[]);
-//! 3. If you access raw storage, where ever you see something like:
-//!      deps.storage.get(CONFIG_KEY).expect("no data stored");
-//!    replace it with:
-//!      deps.with_storage(|store| {
-//!          let data = store.get(CONFIG_KEY).expect("no data stored");
-//!          //...
-//!      });
-//! 4. Anywhere you see query(&deps, ...) you must replace it with query(&mut deps, ...)
-
-use cosmwasm_std::{
-    from_binary, Coin, Decimal, HandleResponse, HandleResult, HumanAddr, InitResponse, StdError,
-};
-use cosmwasm_vm::testing::{
-    handle, init, mock_dependencies, mock_env, query, MockApi, MockQuerier, MockStorage,
-};
-use cosmwasm_vm::Instance;
-use mirror_protocol::mint::{ConfigResponse, HandleMsg, InitMsg, QueryMsg};
-
-// This line will test the output of cargo wasm
-static WASM: &[u8] =
-    include_bytes!("../../../target/wasm32-unknown-unknown/release/mirror_mint.wasm");
-// You can uncomment this line instead to test productionified build from rust-optimizer
-// static WASM: &[u8] = include_bytes!("../contract.wasm");
-
-const DEFAULT_GAS_LIMIT: u64 = 500_000;
-
-pub fn mock_instance(
-    wasm: &[u8],
-    contract_balance: &[Coin],
-) -> Instance<MockStorage, MockApi, MockQuerier> {
-    // TODO: check_wasm is not exported from cosmwasm_vm
-    // let terra_features = features_from_csv("staking,terra");
-    // check_wasm(wasm, &terra_features).unwrap();
-    let deps = mock_dependencies(20, contract_balance);
-    Instance::from_code(wasm, deps, DEFAULT_GAS_LIMIT).unwrap()
-}
-
-static TOKEN_CODE_ID: u64 = 10u64;
-
-#[test]
-fn proper_initialization() {
-    let mut deps = mock_instance(WASM, &[]);
-
-    let msg = InitMsg {
-        owner: HumanAddr::from("owner0000"),
-        oracle: HumanAddr::from("oracle0000"),
-        collector: HumanAddr::from("collector0000"),
-        base_denom: "uusd".to_string(),
-        token_code_id: TOKEN_CODE_ID,
-        protocol_fee_rate: Decimal::percent(1),
-    };
-
-    let env = mock_env("addr0000", &[]);
-
-    // we can just call .unwrap() to assert this was a success
-    let _res: InitResponse = init(&mut deps, env.clone(), msg).unwrap();
-
-    // it worked, let's query the state
-    let res = query(&mut deps, QueryMsg::Config {}).unwrap();
-    let config: ConfigResponse = from_binary(&res).unwrap();
-    assert_eq!("owner0000", config.owner.as_str());
-    assert_eq!("uusd", config.base_denom.to_string());
-    assert_eq!("oracle0000", config.oracle.as_str());
-    assert_eq!(TOKEN_CODE_ID, config.token_code_id);
-}
-
-#[test]
-fn update_config() {
-    let mut deps = mock_instance(WASM, &[]);
-    let msg = InitMsg {
-        owner: HumanAddr::from("owner0000"),
-        oracle: HumanAddr::from("oracle0000"),
-        collector: HumanAddr::from("collector0000"),
-        base_denom: "uusd".to_string(),
-        token_code_id: TOKEN_CODE_ID,
-        protocol_fee_rate: Decimal::percent(1),
-    };
-
-    let env = mock_env("addr0000", &[]);
-
-    // we can just call .unwrap() to assert this was a success
-    let _res: InitResponse = init(&mut deps, env.clone(), msg).unwrap();
-
-    // update owner
-    let env = mock_env("owner0000", &[]);
-    let msg = HandleMsg::UpdateConfig {
-        owner: Some(HumanAddr("owner0001".to_string())),
-        oracle: None,
-        collector: None,
-        token_code_id: Some(100u64),
-        protocol_fee_rate: None,
-    };
-
-    let res: HandleResponse = handle(&mut deps, env, msg).unwrap();
-    assert_eq!(0, res.messages.len());
-
-    // it worked, let's query the state
-    let res = query(&mut deps, QueryMsg::Config {}).unwrap();
-    let config: ConfigResponse = from_binary(&res).unwrap();
-    assert_eq!("owner0001", config.owner.as_str());
-    assert_eq!(100u64, config.token_code_id);
-
-    // Unauthorzied err
-    let env = mock_env("owner0000", &[]);
-    let msg = HandleMsg::UpdateConfig {
-        owner: None,
-        oracle: None,
-        collector: None,
-        token_code_id: None,
-        protocol_fee_rate: None,
-    };
-
-    let res: HandleResult = handle(&mut deps, env, msg);
-    match res.unwrap_err() {
-        StdError::Unauthorized { .. } => {}
-        _ => panic!("Must return unauthorized error"),
-    }
-}
