+++ packages/mirror_protocol/src/mint.rs
@@ -12,6 +12,8 @@ pub struct InitMsg {
     pub owner: HumanAddr,
     pub oracle: HumanAddr,
     pub collector: HumanAddr,
+    pub staking: HumanAddr,
+    pub terraswap_factory: HumanAddr,
     pub base_denom: String,
     pub token_code_id: u64,
     pub protocol_fee_rate: Decimal,
@@ -31,6 +33,7 @@ pub enum HandleMsg {
         owner: Option<HumanAddr>,
         oracle: Option<HumanAddr>,
         collector: Option<HumanAddr>,
+        terraswap_factory: Option<HumanAddr>,
         token_code_id: Option<u64>,
         protocol_fee_rate: Option<Decimal>,
     },
@@ -61,6 +64,7 @@ pub enum HandleMsg {
         collateral: Asset,
         asset_info: AssetInfo,
         collateral_ratio: Decimal,
+        short_params: Option<ShortParams>,
     },
     /// Deposit more collateral
     Deposit {
@@ -76,7 +80,14 @@ pub enum HandleMsg {
     Mint {
         position_idx: Uint128,
         asset: Asset,
+        short_params: Option<ShortParams>,
     },
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+pub struct ShortParams {
+    pub belief_price: Option<Decimal>,
+    pub max_spread: Option<Decimal>,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
@@ -86,6 +97,7 @@ pub enum Cw20HookMsg {
     OpenPosition {
         asset_info: AssetInfo,
         collateral_ratio: Decimal,
+        short_params: Option<ShortParams>,
     },
     /// Deposit more collateral
     Deposit { position_idx: Uint128 },
@@ -121,6 +133,8 @@ pub struct ConfigResponse {
     pub owner: HumanAddr,
     pub oracle: HumanAddr,
     pub collector: HumanAddr,
+    pub staking: HumanAddr,
+    pub terraswap_factory: HumanAddr,
     pub base_denom: String,
     pub token_code_id: u64,
     pub protocol_fee_rate: Decimal,
@@ -143,6 +157,7 @@ pub struct PositionResponse {
     pub owner: HumanAddr,
     pub collateral: Asset,
     pub asset: Asset,
+    pub is_short: bool,
 }
 
 #[derive(Serialize, Deserialize, Clone, PartialEq, JsonSchema, Debug, Default)]
@@ -157,4 +172,7 @@ pub struct NextPositionIdxResponse {
 
 /// We currently take no arguments for migrations
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct MigrateMsg {}
+pub struct MigrateMsg {
+    pub staking: HumanAddr,
+    pub terraswap_factory: HumanAddr,
+}
